Relative links breakthrough!
I was having a problem where my relative links weren't working because of the pretty url file structure in the _site folder.

Example: content/recipes/cake.md > _site/recipes/cake/index.html.

So that adds another level that the relative link, as set in `content/`, can't account for. `<base>` seems to fix this.