Fix empty tags issue and build
Notes JSON file was giving all notes the "notes" tag, which even though it doesn't render on the page, still means the tags array for any note isn't empty. So even if it was empty, the ul would still render and just look like a block of white space.

Anyway, we're defining collections in the config file now anyway, so it doesn't matter.